change env var gate and move some logic to cli kit for others to use
